Print pid_t's as uintmax_t for portability (solaris)
git-svn-id: http://www.varnish-cache.org/svn/trunk@4562 d4fa192b-c00b-0410-8231-f00ffab90ce4

+++ lib/libvarnish/vpf.c
@@ -194,7 +194,7 @@ vpf_write(struct pidfh *pfh)
 		return (-1);
 	}
 
-	error = snprintf(pidstr, sizeof(pidstr), "%u", getpid());
+	error = snprintf(pidstr, sizeof(pidstr), "%ju", (uintmax_t)getpid());
 	assert(error < sizeof pidstr);
 	if (pwrite(fd, pidstr, strlen(pidstr), 0) != (ssize_t)strlen(pidstr)) {
 		error = errno;
